backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
pypdf
langchain-pinecone
python-multipart
tqdm